+++ flexmeasures/data/schemas/sensors.py
@@ -1,0 +1,33 @@
+from marshmallow import Schema, fields
+
+from flexmeasures.data import ma
+from flexmeasures.data.models.time_series import Sensor
+
+
+class SensorSchemaMixin(Schema):
+    """
+    Base sensor schema.
+
+    Here we include all fields which are implemented by timely_beliefs.SensorDBMixin
+    All classes inheriting from timely beliefs sensor don't need to repeat these.
+    In a while, this schema can represent our unified Sensor class.
+
+    When subclassing, also subclass from `ma.SQLAlchemySchema` and add your own DB model class, e.g.:
+
+        class Meta:
+            model = Asset
+    """
+
+    name = ma.auto_field(required=True)
+    unit = ma.auto_field(required=True)
+    timezone = ma.auto_field()
+    event_resolution = fields.TimeDelta(required=True, precision="minutes")
+
+
+class SensorSchema(SensorSchemaMixin, ma.SQLAlchemySchema):
+    """
+    Sensor schema, with validations.
+    """
+
+    class Meta:
+        model = Sensor