+++ flexmeasures/data/schemas/utils.py
@@ -1,0 +1,26 @@
+import click
+import marshmallow as ma
+from marshmallow import ValidationError
+
+
+class MarshmallowClickMixin(click.ParamType):
+    def get_metavar(self, param):
+        return self.__class__.__name__
+
+    def convert(self, value, param, ctx, **kwargs):
+        try:
+            return self.deserialize(value, **kwargs)
+        except ma.exceptions.ValidationError as e:
+            raise click.exceptions.BadParameter(e, ctx=ctx, param=param)
+
+
+class FMValidationError(ValidationError):
+    """
+    Custom validation error class.
+    It differs from the classic validation error by having two
+    attributes, according to the USEF 2015 reference implementation.
+    Subclasses of this error might adjust the `status` attribute accordingly.
+    """
+
+    result = "Rejected"
+    status = "UNPROCESSABLE_ENTITY"